fixed docker for windows

+++ run.sh
@@ -8,7 +8,7 @@ set -e
 
 # Run required migrations
 export FLASK_APP=core/server.py
-rm core/store.sqlite3
+[ -f core/store.sqlite3 ] && rm core/store.sqlite3
 # flask db init -d core/migrations/
 # flask db migrate -m "Initial migration." -d core/migrations/
 flask db upgrade -d core/migrations/
